Перенес весь код из модуля image-preview.js в модуль с формой image-upload-form.js.
